Aggiunti services, eccezioni, configurazioni e integrazione Keycloak:
- Aggiunte classi Service (CarrelloService, OrdineService, RegistrazioneService, VideogiocoService)
- Integrazione con Keycloak per registrazione utenti, logout e claim custom (idUtente)
- Aggiunte eccezioni custom
- Creata configurazione Security con JWT
- Aggiunte classi di supporto (JwtAuthenticationConverter, Utils, ResponseMessage)
- Aggiornato pom.xml con dipendenze Keycloak
- Aggiornato application.yaml con configurazioni Keycloak e JWT

+++ src/main/java/com/example/gamestore/entities/DettaglioCarrello.java
@@ -40,7 +40,7 @@ public class DettaglioCarrello {
 
     @Basic
     @Column(name = "prezzo_unitario")
-    private float prezzoUnitario;
+    private Double prezzoUnitario;
 
 
 }
